WIP: Breaking out container composition tests for reuse
Also removed unused imports in test file
